requirements.txt
flask
requests
tensorflow
numpy
pandas
pillow
matplotlib